requirements.txt
Django==2.1.8
djangorestframework==3.9.4
psycopg2==2.7.5
django-rest-swagger==2.2.0
djangorestframework-gis==0.14
django-filter==2.1.0
python-memcached==1.59